QuestionDataService에 answer을 관리하는 기능 추가 및 서비스 테스트 코드 수정
- update, delete에서만 answer을 관리

+++ src/main/kotlin/com/yearsaday/qna/service/QuestionDataService.kt
@@ -6,6 +6,7 @@ import com.yearsaday.qna.message.QuestionResponse
 import com.yearsaday.qna.message.QuestionUpdateRequest
 import com.yearsaday.qna.repository.QuestionRepository
 import org.springframework.stereotype.Service
+import java.lang.IllegalArgumentException
 
 @Service
 class QuestionDataService(
@@ -26,18 +27,21 @@ class QuestionDataService(
     }
 
     fun delete(id: Int) {
+        if(repository.findById(id).orElseThrow().answers.size > 0)
+            throw IllegalArgumentException()
+
         repository.deleteById(id)
     }
 
     fun update(id: Int, request: QuestionUpdateRequest): QuestionResponse {
         val saved = repository.findById(id).orElseThrow()
-        val update = Question(id, request.sentence)
+        val update = Question(id, request.sentence, saved.answers)
         val result = repository.save(update)
 
         return toQuestionResponse(result)
     }
 
     private fun toQuestionResponse(entity: Question): QuestionResponse {
-        return QuestionResponse(entity.id, entity.sentence)
+        return QuestionResponse(entity.id, entity.sentence, entity.answers)
     }
 }
